Implement settings page and export
Implemented the settings page and export functionality. The about and privacy
pages were linked using embedded Safari browsers. The ID of the device displays
in the settings page and allows the user to copy it right from the app with a
tap. Finally, the export and share functionality was implemented, allowing
the user to export any cached hikes that they might otherwise not be able to
upload.

re #3