Fix inconsistency in connected status
The connectivity status on the record page was very inconsistent, showing red
constantly after recording the first hike. This change moves the code that
starts the monitoring process into the start command for the measurment service
so that the monitoring is properly started each time a hike is recorded.

re #13